Look at checksums before processing a changes file (closes #88374)
This should ensure we don't end up with files in the repositories that
come from different builds.  It's still possible for this to fail if the
files get overwritten between the check and the move; this could be
fixed by first moving files to a staging area, doing the check, then
moving them to their destination.  One step at a time, though.